Create Nikhilam Test Code.py
This is the test code for the Nikhilam Sutra program. The first set of code can be used for manually entering numbers into the num1 and num2 variables to check if you get a valid answer. You can get a recursion error, a solution that is not correct, or a right answer. The second set is a piece of code that will search an extensive range of equations to check how many it can solve. It will return the amount it solved, and you can compare that with the total number of equations by multiplying the two ranges. This piece of code can be modified to fit unique needs.